implement remark as service
